+++ okapi/filters/its/src/main/java/org/w3c/its/ITSTrace.java
@@ -43,7 +43,7 @@ class ITSTrace {
 	String lqIssuesRef;
 	GenericAnnotations lqIssues;
 	String annotatorsRef;
-	Float mtConfidence;
+	Double mtConfidence;
 	GenericAnnotations storageSize;
 	GenericAnnotations disambig;
 	GenericAnnotations lqRating;
